pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "product-kernel"
version = "0.1.0"
description = "Shared FastAPI plumbing (app factory, middlewares, in-memory adapters)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "starlette>=0.36",
  "pydantic>=2.5",
  "SQLAlchemy>=2.0",
  "asyncpg",
  "alembic",
  "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
jwt = ["PyJWT>=2.8.0"]